Add delete dialog on debt

+++ app/src/main/java/com/nurram/project/pencatatkeuangan/view/fragment/debt/DebtFragment.kt
@@ -81,8 +81,19 @@ class DebtFragment : Fragment() {
             if (it1 == "delete") {
                 (parentFragment?.activity as MainActivity).reduceValue("", it.total.toLong())
 
-                viewModel?.deleteDebt(it)
-                Toast.makeText(context, R.string.toast_hapus_berhasil, Toast.LENGTH_SHORT).show()
+                val dialog = AlertDialog.Builder(requireContext())
+                dialog.setTitle(getString(R.string.attention))
+                dialog.setMessage(R.string.delete_record_confirmation)
+                dialog.setCancelable(true)
+                dialog.setPositiveButton("Yes") { _, _ ->
+                    viewModel?.deleteDebt(it)
+                    Toast.makeText(context, R.string.data_success_delete, Toast.LENGTH_SHORT).show()
+                }
+                dialog.setNegativeButton("Cancel") { innerDialog, _ ->
+                    innerDialog.dismiss()
+                }
+
+                dialog.show()
             } else {
                 showAddDataDialog(it)
             }
